Style: removing JsonTest class and refactor: the  the data base classes
