+++ app/src/main/java/com/zoka/moviesapp/data/MoviesContract.java
@@ -13,9 +13,6 @@ public class MoviesContract {
     public static final Uri BASE_CONTENT_URI = Uri.parse("content://" + CONTENT_AUTHORITY);
 
     public static final String PATH_MOVIES = "movies";
-    public static final String PATH_TRAILERS = "trailers";
-    public static final String PATH_REVIEWS = "reviews";
-    public static final String PATH_MOVIES_DETAILS = "movies_details";
 
 
     public static class MoviesEntry implements BaseColumns {
@@ -32,24 +29,6 @@ public class MoviesContract {
         public static final String COLUMN_BACK_DROP_PATH = "back_drop";
         public static final String COLUMN_DATE = "date";
         public static final String COLUMN_RATE = "rate";
-
-
-    }
-
-    public class ReviewsEntry implements BaseColumns {
-        public static final String TABLE_NAME = "reviews";
-        public static final String COLUMN_ID = "id";
-        public static final String COLUMN_URL = "url";
-        public static final String COLUMN_CONTENT = "content";
-        public static final String COLUMN_AUTHOR_NAME = "authorName";
-
-
-    }
-
-    public class TrailersEntry implements BaseColumns {
-        public static final String TABLE_NAME = "trailers";
-        public static final String COLUMN_ID = "id";
-        public static final String COLUMN_KEY = "key";
 
 
     }
