+++ app/src/main/java/com/zoka/moviesapp/data/MoviesDbHelper.java
@@ -9,7 +9,7 @@ import android.database.sqlite.SQLiteOpenHelper;
  */
 
 public class MoviesDbHelper extends SQLiteOpenHelper {
-    private static final int DATABASE_VERSION = 3;
+    private static final int DATABASE_VERSION = 4;
 
     private static final String DATABASE_NAME = "weather.db";
 
@@ -29,26 +29,15 @@ public class MoviesDbHelper extends SQLiteOpenHelper {
                 MoviesContract.MoviesEntry.COLUMN_DESCRIPTION + " TEXT, " +
                 MoviesContract.MoviesEntry.COLUMN_TITLE + " TEXT);";
 
-        final String SQL_CREATE_REVIEWS_TABLE = "CREATE TABLE " + MoviesContract.ReviewsEntry.TABLE_NAME + " (" +
-                MoviesContract.ReviewsEntry.COLUMN_ID + " TEXT  NOT NULL, " +
-                MoviesContract.ReviewsEntry.COLUMN_URL + " TEXT, " +
-                MoviesContract.ReviewsEntry.COLUMN_AUTHOR_NAME + " TEXT UNIQUE NOT NULL, " +
-                MoviesContract.ReviewsEntry.COLUMN_CONTENT + " TEXT);";
 
-        final String SQL_CREATE_TRAILERS_TABLE = "CREATE TABLE " + MoviesContract.TrailersEntry.TABLE_NAME + " (" +
-                MoviesContract.TrailersEntry.COLUMN_ID + " TEXT  NOT NULL, " +
-                MoviesContract.TrailersEntry.COLUMN_KEY + " TEXT UNIQUE NOT NULL);";
 
         db.execSQL(SQL_CREATE_MOVIE_TABLE);
-        db.execSQL(SQL_CREATE_REVIEWS_TABLE);
-        db.execSQL(SQL_CREATE_TRAILERS_TABLE);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion) {
         sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + MoviesContract.MoviesEntry.TABLE_NAME);
-        sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + MoviesContract.ReviewsEntry.TABLE_NAME);
-        sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + MoviesContract.TrailersEntry.TABLE_NAME);
+
         onCreate(sqLiteDatabase);
     }
 }
